Move firmware update classes to internal package
Move firmware update related classes from public to internal package to better encapsulate the API. Classes moved: FirmwareFile, ZWaveFirmwareUpdateSession, and ZWaveLocalFirmwareProvider. Idea is to reduce differences between manual firmware update and using the remote firmware repository.

Also simplifies FirmwareFileContainer by removing the unused format field that was not being utilized downstream.
