+++ src/main/java/org/openhab/binding/zwave/handler/ZWaveThingHandler.java
@@ -42,12 +42,12 @@ import java.util.stream.Stream;
 import org.eclipse.jdt.annotation.Nullable;
 import org.openhab.binding.zwave.ZWaveBindingConstants;
 import org.openhab.binding.zwave.actions.ZWaveThingActions;
-import org.openhab.binding.zwave.firmwareupdate.FirmwareFile;
-import org.openhab.binding.zwave.firmwareupdate.ZWaveFirmwareUpdateSession;
-import org.openhab.binding.zwave.firmwareupdate.ZWaveFirmwareUpdateSession.FirmwareUpdateEvent;
 import org.openhab.binding.zwave.handler.ZWaveThingChannel.DataType;
 import org.openhab.binding.zwave.internal.ZWaveConfigProvider;
 import org.openhab.binding.zwave.internal.ZWaveProduct;
+import org.openhab.binding.zwave.internal.firmwareupdate.FirmwareFile;
+import org.openhab.binding.zwave.internal.firmwareupdate.ZWaveFirmwareUpdateSession;
+import org.openhab.binding.zwave.internal.firmwareupdate.ZWaveFirmwareUpdateSession.FirmwareUpdateEvent;
 import org.openhab.binding.zwave.internal.protocol.ZWaveAssociation;
 import org.openhab.binding.zwave.internal.protocol.ZWaveAssociationGroup;
 import org.openhab.binding.zwave.internal.protocol.ZWaveConfigurationParameter;
