+++ src/main/java/org/openhab/binding/zwave/internal/protocol/commandclass/ZWaveFirmwareUpdateCommandClass.java
@@ -17,7 +17,7 @@ import java.util.Arrays;
 
 import org.eclipse.jdt.annotation.NonNullByDefault;
 import org.eclipse.jdt.annotation.Nullable;
-import org.openhab.binding.zwave.firmwareupdate.ZWaveFirmwareUpdateSession.FirmwareUpdateEvent;
+import org.openhab.binding.zwave.internal.firmwareupdate.ZWaveFirmwareUpdateSession.FirmwareUpdateEvent;
 import org.openhab.binding.zwave.internal.protocol.ZWaveCommandClassPayload;
 import org.openhab.binding.zwave.internal.protocol.ZWaveController;
 import org.openhab.binding.zwave.internal.protocol.ZWaveEndpoint;
