+++ src/test/java/org/openhab/binding/zwave/handler/ZWaveThingHandlerTest.java
@@ -33,7 +33,7 @@ import org.mockito.ArgumentCaptor;
 import org.mockito.ArgumentMatchers;
 import org.mockito.Mockito;
 import org.openhab.binding.zwave.ZWaveBindingConstants;
-import org.openhab.binding.zwave.firmwareupdate.ZWaveFirmwareUpdateSession;
+import org.openhab.binding.zwave.internal.firmwareupdate.ZWaveFirmwareUpdateSession;
 import org.openhab.binding.zwave.internal.protocol.ZWaveAssociationGroup;
 import org.openhab.binding.zwave.internal.protocol.ZWaveController;
 import org.openhab.binding.zwave.internal.protocol.ZWaveDeviceClass;
