+++ src/test/java/org/openhab/binding/zwave/internal/firmwareupdate/ZWaveFirmwareUpdateSessionTest.java
@@ -10,7 +10,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package org.openhab.binding.zwave.firmwareupdate;
+package org.openhab.binding.zwave.internal.firmwareupdate;
 
 import static org.junit.jupiter.api.Assertions.*;
